remove qint64 from  header

+++ SerialPrograms/Source/CommonFramework/Tools/FileDownloader.cpp
@@ -98,7 +98,7 @@ void download_file_to_disk(
     Logger& logger, 
     const std::string& url, 
     const std::string& file_path,
-    qint64 expected_size,
+    uint64_t expected_size,
     std::function<void(uint64_t bytes_done, uint64_t total_bytes)> progress_callback
 ){
 //    cout << "download_file()" << endl;
@@ -134,9 +134,8 @@ void download_file_to_disk(
             }
             
             // Use expected_size if the network doesn't provide one
-            qint64 total = (bytesTotal > 0) ? bytesTotal : expected_size;
+            uint64_t total = (bytesTotal > 0) ? (uint64_t)bytesTotal : expected_size;
 
-            // int percentage_progress = (total > 0) ? static_cast<int>((bytesReceived * 100) / total) : 0;
             progress_callback(bytesReceived, total);
         }
     );
